Add effects system
This allows for buildings to specify effects, which modify values somewhere else in the game. The only effect implemented right now is resourceProductionBase, which increases the base production of a resource.
In the future there will be more effects, and resources will also have the ability to make effects.
The planned effects to add in the future are:
- resourceProductionRatio
- eventRewardRatio
- eventChanceRatio