Added base linked property values and improved resource tooltips
The addition of base linked property values also improves the way linked properties are defined. GameObjects are now constructed with a GameObjectConfig which includes the base values in config.linkablePropertyBaseValues. GameObject now automatically populates linkablePropertyBaseValues with default values before adding the values from the config. It also now creates the deriveds for linkable properties automatically, so linked properties are now just defined like "production = $derived(0)" and are later updated with the proper derived.

The resource table is now a grid to more easily control column spacing. The numeric columns are now large enough to support values up to 1e99 when using Eternal Notations AD Mixed Scientific notation, which will use in the future once it releases as an NPM package.

The maxAmount and production columns in the resource table use a new LinkTree component, which shows a breakdown of how all the links effect the final property value.